+++ gnuplot_manager/demo.py
@@ -45,62 +45,70 @@ def main():
 
     print('\n***Demo of the gnuplot_manager package***\n')
     
-    print('* -> In this demo, gnuplot output is silenced\n')
+    print('In this demo, gnuplot output is silenced\n')
     p2d = new_plot(xpos=600, ypos=100, width=600, height=450,
                    plot_type='2D', title='2D PLOT WINDOW DEMO', redirect_output=None)        
     p3d = new_plot(xpos=680, ypos=200, width=600, height=450,
                    plot_type='3D', title='3D PLOT WINDOW DEMO', redirect_output=None)    
     
-    print('* -> Plotting a parabola from data\n')
+    print(' 1) -> Plotting a parabola from data\n')
     plot2d(p2d, x, x*x, 'parabola')
     wait()
 
-    print('* -> Plotting a parabola as a function\n')
+    print(' 2) -> Plotting a parabola as a function\n')
     plot_function(p2d, 'x**2', replot=True)
     wait()
     
-    print('* -> Changing the scale limits\n')
+    print(' 3) -> Changing the scale limits\n')
     plot_set(p2d, xmin=1, xmax=5, ymin=1, ymax=25, replot=True)
     wait()
 
-    print('* -> Set logarithmic x and y axes\n')
+    print(' 4) -> Setting logarithmic x and y axes\n')
     plot_set(p2d, logx=True, logy=True, replot=True)
     wait()
 
-    print('* -> Adding a label to the plot\n')
+    print(' 5) -> Resetting ranges and scales\n')
+    plot_set(p2d, resetx=True, resety=True, logx=False, logy=False, replot=True)
+    wait()    
+
+    print(' 6) -> Adding a label to the plot\n')
     plot_label(p2d, x=20, y=20, label='THIS IS A LABEL', replot=True)
     wait()
 
-    print('* -> Removing the label\n')
+    print(' 7) -> Removing the label\n')
     plot_label(p2d, label=None, erase=True, replot=True)
     wait()
 
-    print('* -> Plotting a 3D curve from data\n')
+    print(' 8) -> Plotting another curve with volatile data\n')
+    plot2d(p2d, x, x*x*x/100, 'x^3/100 (volatile)', volatile=True, replot=True)
+    wait()            
+
+    print(' 9) -> Plotting a 3D curve from data\n')
     plot3d(p3d, x, x, 2*x*x, '3D curve')
     wait()
 
-    print('* -> Plotting a 3D surface as a function\n')
+    print('10) -> Plotting a 3D surface as a function\n')
     plot_function(p3d, 'x**2+y**2', replot=True)
     wait()
 
-    print('* -> Setting logarithmic z axis\n')
+    print('11) -> Setting logarithmic z axis\n')
     plot_set(p3d, logz=True, replot=True)
     wait()
 
-    print('* -> Listing the plots\n')
+    print('12) -> Listing the plots\n')
     plot_list()
     wait()
 
-    print('* -> Closing the 2d plot, removing data files\n')
-    plot_close(p2d, purge=True)
+    print('13) -> Closing the 2d plot\n')
+    plot_close(p2d)
     wait()
 
-    print('* -> Listing the plots again')
+    print('14) -> Listing the plots again')
     plot_list()
     wait()    
 
-    print('* -> Closing all the plots, removing data files\n')
-    plot_close_all(purge=True)
+    print('15) -> Closing all the plots, removing files and directories\n')
+    plot_close_all()
 
     print('***End of demo***')