+++ gnuplot_manager/errors.py
@@ -34,9 +34,13 @@ ERROR_NO_DATA         = (22, 'no data to plot')
 ERROR_INVALID_STYLE   = (23, 'invalid plot style')
 ERROR_WRONG_TYPE      = (24, 'invalid data type')
 ERROR_REPLOT_VOLATILE = (25, 'replot impossible with volatile curves')
+ERROR_LOGSET_VOLATILE = (26, 'logscale toggle impossible with volatile curves')
 
-# Error about parameters inconsistency (e.g. xmin>xmax)
+# Errors about parameters inconsistency (e.g. xmin>xmax)
 ERROR_PLOT_PARAMETERS = (30, 'parameters error')
+
+# Errors about file operations
+ERROR_FILE            = (40, 'file error')